ldap: Instance Name
This addresses an issue where if you have more than one LDAP instance
running it will show "Active Directory or LDAP" for all instances making it
hard to differentiate the instances. This updates the `getName()` function
to grab the name from the config if available otherwise defaults to the
default name.