Auth-OAuth2: Add Plugin Instance Fix
When adding a new instance of MicrosoftEmailOauth2 the core validates
data using default plugin config form instead of the custom one.

This commit addresses the issue temporarily by injecting resource owner
url for validation purpose only. In the upcoming version of core we will
accept either a form or vars - making it possible to support older
plugins until end of life is reached.